Split out working to reduce CPU spikes in a small amount of time
