+++ app/jobs/working_days/update_assessments_submission_to_started_job.rb
@@ -1,0 +1,30 @@
+# frozen_string_literal: true
+
+class WorkingDays::UpdateAssessmentsSubmissionToStartedJob < ApplicationJob
+  def perform
+    Assessment
+      .joins(:application_form)
+      .includes(:application_form)
+      .where.not(application_form: { submitted_at: nil })
+      .where.not(started_at: nil)
+      .find_each do |assessment|
+        assessment.update!(
+          working_days_between_submitted_and_started:
+            calendar.business_days_between(
+              assessment.application_form.submitted_at,
+              assessment.started_at,
+            ),
+        )
+      end
+  end
+
+  private
+
+  def calendar
+    @calendar ||=
+      Business::Calendar.new(
+        holidays:
+          DfE::ReferenceData::BankHolidays::BANK_HOLIDAYS.all.map(&:date),
+      )
+  end
+end
